+++ src/main/java/http/request/Protocol.java
@@ -1,4 +1,4 @@
-package http;
+package http.request;
 
 import java.util.Objects;
