+++ src/test/java/http/RequestMethodGetTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.RequestMethodGet;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
