+++ src/test/java/http/RequestParametersTest.java
@@ -1,11 +1,9 @@
 package http;
 
+import http.request.RequestParameters;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
